Refactor replacement operations
Added some benchmarks for various way of running the code with some
worst case scenarios. Based off of those the code has been modified to
be a bit faster.

The benchmarks hints to that there's no way that's a lot faster than the
other, so while I've optimized according to what is faster here I'm not
sure it'll yield anything in real world scenarios.